Commit - Constraints redefined, made other changes
Made some adjustments to code and enabled foreign key constraint

+++ main.py
@@ -1,4 +1,5 @@
-from DB_Models.Wishlist.wishlist import Wishlist
+
+from sqlalchemy.sql.elements import Null
 from sqlalchemy.sql.expression import null
 import DB_Models
 from Models import *
@@ -10,6 +11,8 @@ from DB_Models import Orders, User, Restaurant, Inventory, Wishlist
 from DB_Models.Orders.orders import *
 from DB_Models.Inventory.inventory import *
 from DB_Models.Menu.menu import *
+from DB_Models.Wishlist.wishlist import *
+from DB_Models.User.user import *
 from Models.Order.order import *
 from Models.Menu.menu import *
 from Models.User.user import *
@@ -19,7 +22,11 @@ from Models.Wishlist.wishlist import *
 from DB_Models.database import SessionLocal, engine
 from passlib.context import CryptContext
 from datetime import datetime
+from sqlalchemy import event
+from sqlalchemy.engine import Engine
+from sqlite3 import Connection as SQLite3Connection
 rest = FastAPI()
+
 
 pwd_context = CryptContext(
     schemes=[
@@ -40,7 +47,15 @@ def get_db():
 DB_Models.database.Base.metadata.create_all(bind=engine)
 
 
+@event.listens_for(Engine, "connect")
+def _set_sqlite_pragma(dbapi_connection, connection_record):
+    if isinstance(dbapi_connection, SQLite3Connection):
+        cursor = dbapi_connection.cursor()
+        cursor.execute("PRAGMA foreign_keys=ON;")
+        cursor.close()
+
 # Endpoints for User Details
+
 
 @rest.get("/user/all")
 def user_details(db: Session = Depends(get_db)):
@@ -80,12 +95,14 @@ def create_user(detail_request: UserDetails, db: Session = Depends(get_db)):
 @rest.delete("/user/{id}")
 def del_user_by_id(id: int, db: Session = Depends(get_db)):
 
-    db.query(User).filter(User.id == id).delete()
-    db.commit()
-
-    return {
-        "code": "success",
-        "message": "User details deleted from the database"}
+    if db.query(User).filter(User.id == id):
+        db.query(User).filter(User.id == id).delete()
+        db.commit()
+        return {
+            "code": "success",
+            "message": "User details deleted from the database"}
+    else:
+        return "No such User ID"
 
 
 @rest.put("/user/{id}")
@@ -97,7 +114,10 @@ def update_user(id: int, item: UserEdit, db: Session = Depends(get_db)):
 
     db.commit()
     data = db.query(User).filter(User.id == id).first()
-    return data
+    if data:
+        return data
+    else:
+        return "No such used id exists"
 
 
 @rest.get("/user/totalspent/{id}")
@@ -118,7 +138,7 @@ def total_spent(id: int, db: Session = Depends(get_db)):
 @rest.get("/user/orderhistory/{id}")
 def order_history(id: int, db: Session = Depends(get_db)):
 
-    history = [("User Id", "Date")]
+    history = [("Order Id", "Date")]
     values = [(i[0], i[1].strftime('%m/%d/%Y'))
               for i in db.query(Orders.id, Orders.date).filter(Orders.userid == id).all()]
     history.extend(values)
@@ -168,12 +188,14 @@ def create_rest(detail_request: RestDetails, db: Session = Depends(get_db)):
 @rest.delete("/rest/{id}")
 def del_rest_by_id(id: int, db: Session = Depends(get_db)):
 
-    db.query(Restaurant).filter(Restaurant.id == id).delete()
-    db.commit()
-
-    return {
-        "code": "success",
-        "message": "Restaurant details deleted from the database"}
+    if db.query(Restaurant).filter(Restaurant.id == id):
+        db.query(Restaurant).filter(Restaurant.id == id).delete()
+        db.commit()
+        return {
+            "code": "success",
+            "message": "Restaurant details deleted from the database"}
+    else:
+        return "No such Restaurant ID"
 
 
 @rest.put("/rest/{id}")
@@ -186,7 +208,10 @@ def update_rest(id: int, item: RestEdit, db: Session = Depends(get_db)):
 
     db.commit()
     data = db.query(Restaurant).filter(Restaurant.id == id).first()
-    return data
+    if data:
+        return data
+    else:
+        return "No such restaurant id exists"
 
 # Endpoints for Menu Category Details
 
@@ -204,30 +229,37 @@ def category_by_id(id: int, db: Session = Depends(get_db)):
     return {"data": data}
 
 
-@rest.post("/categ/", response_model=CategDetails)
+@rest.post("/categ/")
 def create_category(
         detail_request: CategDetails,
         db: Session = Depends(get_db)):
 
-    post = Menu_Category()
-    post.name = detail_request.name
-    post.description = detail_request.description
-    post.restid = detail_request.restid
+    if not db.query(Menu_Category).filter(
+            Menu_Category.name == detail_request.name).filter(
+            Menu_Category.restid == detail_request.restid).first():
+        post = Menu_Category()
+        post.name = detail_request.name
+        post.description = detail_request.description
+        post.restid = detail_request.restid
 
-    db.add(post)
-    db.commit()
-    return detail_request
+        db.add(post)
+        db.commit()
+        return detail_request
+    else:
+        return "Data already exists for given inputs"
 
 
 @rest.delete("/categ/{id}")
 def del_category_by_id(id: int, db: Session = Depends(get_db)):
 
-    db.query(Menu_Category).filter(Menu_Category.id == id).delete()
-    db.commit()
-
-    return {
-        "code": "success",
-        "message": "Menu details deleted from the database"}
+    if db.query(Menu_Category).filter(Menu_Category.id == id):
+        db.query(Menu_Category).filter(Menu_Category.id == id).delete()
+        db.commit()
+        return {
+            "code": "success",
+            "message": "Menu Category details deleted from the database"}
+    else:
+        return "No such Category ID"
 
 
 @rest.put("/categ/{id}")
@@ -240,7 +272,10 @@ def update_category(id: int, item: CategEdit, db: Session = Depends(get_db)):
 
     db.commit()
     data = db.query(Menu_Category).filter(Menu_Category.id == id).first()
-    return data
+    if data:
+        return data
+    else:
+        return "No such category id exists"
 
 
 # Endpoints for Menu Item Details
@@ -265,31 +300,39 @@ def item_by_id(id: int, db: Session = Depends(get_db)):
     return {"data": data}
 
 
-@rest.post("/item/", response_model=ItemDetails)
+@rest.post("/item/")
 def create_item(detail_request: ItemDetails, db: Session = Depends(get_db)):
 
-    post = Menu_Item()
-    post.name = detail_request.name
-    post.description = detail_request.description
-    post.categid = detail_request.categid
-    post.cost = detail_request.cost
-    post.gluten_free = detail_request.gluten_free
-    post.vegetarian = detail_request.vegetarian
+    if not db.query(Menu_Item).filter(
+            Menu_Item.name == detail_request.name).filter(
+            Menu_Item.categid == detail_request.categid).first():
 
-    db.add(post)
-    db.commit()
-    return detail_request
+        post = Menu_Item()
+        post.name = detail_request.name
+        post.description = detail_request.description
+        post.categid = detail_request.categid
+        post.cost = detail_request.cost
+        post.gluten_free = detail_request.gluten_free
+        post.vegetarian = detail_request.vegetarian
+
+        db.add(post)
+        db.commit()
+        return detail_request
+    else:
+        return "Data already exists for the given inputs"
 
 
 @rest.delete("/item/{id}")
 def del_item_by_id(id: int, db: Session = Depends(get_db)):
 
-    db.query(Menu_Item).filter(Menu_Item.id == id).delete()
-    db.commit()
-
-    return {
-        "code": "success",
-        "message": "Item details deleted from the database"}
+    if db.query(Menu_Item).filter(Menu_Item.id == id):
+        db.query(Menu_Item).filter(Menu_Item.id == id).delete()
+        db.commit()
+        return {
+            "code": "success",
+            "message": "Menu Item details deleted from the database"}
+    else:
+        return "No such Menu Item ID"
 
 
 @rest.put("/item/{id}")
@@ -301,26 +344,29 @@ def update_item(id: int, item: ItemEdit, db: Session = Depends(get_db)):
 
     db.commit()
     data = db.query(Menu_Item).filter(Menu_Item.id == id).first()
-    return data
+    if data:
+        return data
+    else:
+        return "No such item id exists"
 
 
 # Endpoints for Order Details
 
 @rest.get("/order/all")
-def item_details(db: Session = Depends(get_db)):
+def order_details(db: Session = Depends(get_db)):
     data = db.query(Orders).all()
     return {"data": data}
 
 
 @rest.get("/order/{id}")
-def item_by_name(id: int, db: Session = Depends(get_db)):
+def order_by_name(id: int, db: Session = Depends(get_db)):
 
     data = db.query(Orders).filter(Orders.id == id).all()
     return {"data": data}
 
 
 @rest.post("/order/")
-def create_item(detail_request: OrderDetails, db: Session = Depends(get_db)):
+def create_order(detail_request: OrderDetails, db: Session = Depends(get_db)):
 
     post = Orders()
     post.userid = detail_request.userid
@@ -330,27 +376,43 @@ def create_item(detail_request: OrderDetails, db: Session = Depends(get_db)):
     post.date = detail_request.date
     post.itemid = detail_request.itemid
 
-    if(int(db.query(Inventory.quantity).filter(Inventory.itemid == detail_request.itemid and Inventory.restaurantid == detail_request.restid).first()[0]) > detail_request.quantity):
-        db.add(post)
-        db.commit()
-        return detail_request
+    if(db.query(Inventory.quantity).filter(Inventory.itemid == detail_request.itemid).first() is not None and
+       db.query(Inventory.quantity).filter(Inventory.restaurantid == detail_request.restid).first() is not None):
+
+        if(db.query(Inventory.quantity).filter(Inventory.itemid == detail_request.itemid)
+           .filter(Inventory.restaurantid == detail_request.restid).first()[0] >= detail_request.quantity):
+            db.add(post)
+
+            newquantity = db.query(
+                Inventory.quantity).filter(
+                Inventory.itemid == detail_request.itemid).filter(
+                Inventory.restaurantid == detail_request.restid).first()[0] - detail_request.quantity
+            db.query(Inventory).filter(Inventory.itemid == detail_request.itemid).filter(
+                Inventory.restaurantid == detail_request.restid).update({"quantity": newquantity})
+            db.commit()
+            return detail_request
+        else:
+            return "Not enough inventory to make that order"
+
     else:
-        return "Not enough inventory to make that order !"
+        return "Not enough inventory to make that order"
 
 
 @rest.delete("/order/{id}")
-def del_item_by_id(id: int, db: Session = Depends(get_db)):
+def del_order_by_id(id: int, db: Session = Depends(get_db)):
 
-    db.query(Orders).filter(Orders.id == id).delete()
-    db.commit()
-
-    return {
-        "code": "success",
-        "message": "Item details deleted from the database"}
+    if db.query(Orders).filter(Orders.id == id):
+        db.query(Orders).filter(Orders.id == id).delete()
+        db.commit()
+        return {
+            "Code": "Success",
+            "Message": "Order details deleted from the database"}
+    else:
+        return "No such Order ID"
 
 
 @rest.put("/order/{id}")
-def update_item(id: int, item: OrderEdit, db: Session = Depends(get_db)):
+def update_order(id: int, item: OrderEdit, db: Session = Depends(get_db)):
 
     update_item_encoded = {i: jsonable_encoder(item)[i] for i in jsonable_encoder(
         item) if jsonable_encoder(item)[i] is not None}
@@ -358,7 +420,11 @@ def update_item(id: int, item: OrderEdit, db: Session = Depends(get_db)):
 
     db.commit()
     data = db.query(Orders).filter(Orders.id == id).first()
-    return data
+    if data:
+        return data
+    else:
+        return "No such order id exists"
+
 
 # Endpoints for Inventory Details
 
@@ -376,31 +442,40 @@ def inventory_by_id(id: int, db: Session = Depends(get_db)):
     return {"data": data}
 
 
-@rest.post("/inventory/", response_model=InventoryDetails)
+@rest.post("/inventory/")
 def create_inventory(
         detail_request: InventoryDetails,
         db: Session = Depends(get_db)):
 
-    post = Inventory()
-    post.name = detail_request.name
-    post.quantity = detail_request.quantity
-    post.itemid = detail_request.itemid
-    post.restaurantid = detail_request.restaurantid
+    if not db.query(Inventory).filter(
+            Inventory.itemid == detail_request.itemid).filter(
+            Inventory.restaurantid == detail_request.restaurantid).first():
 
-    db.add(post)
-    db.commit()
-    return detail_request
+        post = Inventory()
+        # post.name = detail_request.name
+        post.quantity = detail_request.quantity
+        post.itemid = detail_request.itemid
+        post.restaurantid = detail_request.restaurantid
+
+        db.add(post)
+        db.commit()
+        return detail_request
+
+    else:
+        return "Data already exists for given inputs"
 
 
 @rest.delete("/inventory/{id}")
 def del_inventory_by_id(id: int, db: Session = Depends(get_db)):
 
-    db.query(Inventory).filter(Inventory.id == id).delete()
-    db.commit()
-
-    return {
-        "code": "success",
-        "message": "Inventory details deleted from the database"}
+    if db.query(Inventory).filter(Inventory.id == id):
+        db.query(Inventory).filter(Inventory.id == id).delete()
+        db.commit()
+        return {
+            "Code": "Success",
+            "Message": "Inventory details deleted from the database"}
+    else:
+        return "No such Inventory ID"
 
 
 @rest.put("/inventory/{id}")
@@ -415,7 +490,11 @@ def update_inventory(
 
     db.commit()
     data = db.query(Inventory).filter(Inventory.id == id).first()
-    return data
+    if data:
+        return data
+    else:
+        return "No such inventory id exists"
+
 
 # Endpoints for Wishlist Details
 
@@ -452,12 +531,14 @@ def create_wishlist(
 @rest.delete("/wishlist/{id}")
 def del_wishlist_by_id(id: int, db: Session = Depends(get_db)):
 
-    db.query(Wishlist).filter(Wishlist.id == id).delete()
-    db.commit()
-
-    return {
-        "code": "success",
-        "message": "Wishlist details deleted from the database"}
+    if db.query(Wishlist).filter(Wishlist.id == id):
+        db.query(Wishlist).filter(Wishlist.id == id).delete()
+        db.commit()
+        return {
+            "Code": "Success",
+            "Message": "Wishlist details deleted from the database"}
+    else:
+        return "No such Wish list ID"
 
 
 @rest.put("/wishlist/{id}")
@@ -472,4 +553,7 @@ def update_wishlist(
 
     db.commit()
     data = db.query(Wishlist).filter(Wishlist.id == id).first()
-    return data
+    if data:
+        return data
+    else:
+        return "No such wishlist id exists"
